% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycdbdataframe.R
\docType{data}
\name{nycdb}
\alias{nycdb}
\title{NYC ED Level election reaults 2013-2019}
\format{A data frame with wtih 17 columns}
\usage{
nycdb
}
\description{
data frame containing nyc ed level election results
}
\keyword{datasets}
